+++ src/main/kotlin/no/nav/helse/modell/risiko/RisikovurderingDto.kt
@@ -6,7 +6,7 @@ import java.util.*
 class Risikovurdering private constructor(
     private val vedtaksperiodeId: UUID,
     private val opprettet: LocalDateTime,
-    private val samletScore: Int,
+    private val samletScore: Double,
     private val faresignaler: List<String>,
     private val arbeidsuførhetvurdering: List<String>,
     private val ufullstendig: Boolean
@@ -15,7 +15,7 @@ class Risikovurdering private constructor(
         fun restore(
             vedtaksperiodeId: UUID,
             opprettet: LocalDateTime,
-            samletScore: Int,
+            samletScore: Double,
             faresignaler: List<String>,
             arbeidsuførhetvurdering: List<String>,
             ufullstendig: Boolean
@@ -37,7 +37,7 @@ class Risikovurdering private constructor(
 class RisikovurderingDto(
     val vedtaksperiodeId: UUID,
     val opprettet: LocalDateTime,
-    val samletScore: Int,
+    val samletScore: Double,
     val faresignaler: List<String>,
     val arbeidsuførhetvurdering: List<String>,
     val ufullstendig: Boolean
